Feature/bb2 48 merging separating receipts (#88)
* add function getReceipt

* add changeIfMerged function

* add ReceiptShortView struct in structures.hpp

* add getReceipts function

* deleate build and vscode folder

* add mergedreceipts field to receiptHistory

* delete build folder

* add mergeReceipt function

* add unmergeReceipt function

* deleating a mistake in receiptOperations.hpp

* adding merging and unmerging receipts


* correcting rebase errors

* making variables const

* limiting memory consumption

* correcting mistakes  (wrong spelling)

---------

+++ src/core/structures.hpp
@@ -7,24 +7,24 @@ namespace core {
 
 struct User {
   std::string userID;
-  float price;
+  double price;
 };
 
 struct Item {
   std::string whoBuy;
   std::string itemName;
-  float price;
-  float amount;
+  double price;
+  double amount;
   bool paid;
 };
 
 struct Receipt {
-  std::string author;
-  std::vector<std::string> usersIncluded;
-  std::string receiptName;
-  std::string date;
-  std::vector<bakcyl::core::Item> items;
-  std::vector<std::string> mergedReceipts;
+    std::string author;
+    std::vector<std::string> usersIncluded;
+    std::string receiptName;
+    std::string date;
+    std::vector<bakcyl::core::Item> items;
+    std::vector<std::string> mergedReceipts;
 };
 
 struct ReceiptShortView {
